feat(app): balancefetcher component
